requirements.txt
opencv-python>=4.7,<5
numpy>=1.23,<2
mediapipe>=0.10,<1
pygame>=2.5,<3
scipy>=1.10,<2
PyYAML>=6,<7
